requirements.txt
# Core dependencies (existing)
torch==2.5.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.5.0
scipy==1.13.0
matplotlib==3.9.0
seaborn==0.13.2
PyYAML==6.0.1
tqdm==4.66.0
joblib==1.4.0
Pillow==10.3.0

# API dependencies (new)
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-multipart==0.0.9
aiofiles==24.1.0
python-dotenv==1.0.1

# Caching (optional for advanced deployment)
redis==5.0.0
hiredis==2.3.0

# Database (optional for logging)
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.0
httpx==0.27.0

# Monitoring
python-json-logger==2.0.7
